+++ scripts/postinstall.ts
@@ -6,16 +6,16 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import { execSync } from 'child_process';
-import { existsSync } from 'fs';
-import { resolve } from 'path';
+import { exec } from 'node:child_process';
+import { existsSync } from 'node:fs';
+import { resolve } from 'node:path';
 
 import { build } from './esbuild-renderer';
 
 if (!existsSync(`scripts/nordic-publish.js`)) {
     console.log('Building nordic-publish.js');
     const command = `npx esbuild scripts/nordic-publish.ts --bundle --outfile=scripts/nordic-publish.js --platform=node --log-level=warning --minify`;
-    execSync(command, { encoding: 'utf-8' });
+    exec(command, { encoding: 'utf-8' });
 }
 
 if (!existsSync(`dist/bootstrap.css`)) {
@@ -33,5 +33,5 @@ if (!existsSync(`typings/generated/src/index.d.ts`)) {
     const root = resolve(__dirname, '..');
     process.chdir(root);
 
-    execSync('npm run generate-types', { encoding: 'utf-8' });
+    exec('npm run generate-types', { encoding: 'utf-8' });
 }